Massive, terrible test fix pass
At time of committing, the tests (usually) pass.

Long story long: this code is full of whitespace intricacies which my whitespace
cleaner ran in to trouble with. Second cool part: several of the tests are
non-deterministic. At time of this commit, the tests have passed ten times in a
row. I *think* I've nailed down one of the non-deterministic tests to be more
reliable. We'll see what CI thinks.